Week-5/multi-agent-support-system/requirements.txt
python-dotenv
langchain
langgraph
langchain-aws
langchain-tavily
langchain-community
langchain-text-splitters
langchain-chroma
pypdf
tavily-python
chromadb
